# openapi/beanflow-v1.yaml
openapi: 3.1.0
info:
  title: BeanFlow API
  version: 0.2.0
  x-beanflow-contract-status: target
  x-beanflow-contract-date: 2026-08-12
  description: |
    Pre-release target contract for BeanFlow discovery, ordering, payment,
    pickup, loyalty, settlement, and dispute workflows. An operation in this
    file is not evidence that it is deployed. The machine-readable contract for
    currently implemented public controller mappings is beanflow-v1-runtime.yaml.

    All monetary values are integer KRW. All timestamps are ISO-8601 date-time
    values with an offset or UTC designator. A 202 response means processing or
    reconciliation is still in progress and must not be interpreted as success.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /support/cases/{caseId}/timeline:
    get:
      tags: [Support]
      operationId: listSupportCaseTimeline
      summary: List a bounded cross-context timeline for one SupportCase
      description: |
        Merges closed, masked Support and owner facts using the stable tuple
        occurredAt DESC, source rank ASC, itemId DESC. The signed cursor is bound
        to the Case and canonical source/type filters for 15 minutes. A required
        owner failure returns 503; it is never converted to a partial or empty page.
      parameters:
        - $ref: "#/components/parameters/SupportCaseId"
        - $ref: "#/components/parameters/SupportTimelineSources"
        - $ref: "#/components/parameters/SupportTimelineTypes"
        - $ref: "#/components/parameters/SupportTimelineCursor"
        - $ref: "#/components/parameters/SupportTimelineLimit"
      responses:
        "200":
          description: Globally ordered masked transaction facts
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTimelinePage"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/DependencyUnavailable" }
  /support/orders/{orderId}/timeline:
    get:
      tags: [Support]
      operationId: listSupportOrderTimeline
      summary: List owner facts for an Order linked to one SupportCase
      description: |
        Requires both Support Case/Order read permissions, current assignment and
        an active RELATED_ORDER link. SUPPORT is not a valid explicit source for
        this owner-only endpoint. Authorization is rechecked before response release.
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - $ref: "#/components/parameters/SupportOrderTimelineCaseId"
        - $ref: "#/components/parameters/SupportTimelineSources"
        - $ref: "#/components/parameters/SupportTimelineTypes"
        - $ref: "#/components/parameters/SupportTimelineCursor"
        - $ref: "#/components/parameters/SupportTimelineLimit"
      responses:
        "200":
          description: Globally ordered masked owner facts
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTimelinePage"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/DependencyUnavailable" }
...
  /support/orders/{orderId}/post-acceptance-resolutions:
    post:
      tags: [Support]
      operationId: createPostAcceptanceResolution
      summary: Plan an approved post-acceptance resolution without rewriting Order facts
      description: |
        Consumes only the exact approved S60 POST_ACCEPTANCE_RESOLUTION revision.
        The canonical action digest uses support-command-payload/v1 operation
        POST_ACCEPTANCE_RESOLUTION_ACTION_V1 with ordered orderId, outcome,
        responsibility, cashRefundKrw, restorePoints, restoreCoupon,
        settlementAdjustmentKrw and evidenceDigest fields. No owner command runs
        until the separate execution operation consumes the revision.
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostAcceptanceResolutionRequest"
      responses:
        "201":
          description: Immutable resolution plan created or the first exact response replayed
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAcceptanceResolutionResource"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/DependencyUnavailable" }
  /support/post-acceptance-resolutions/{resolutionId}:
    get:
      tags: [Support]
      operationId: getPostAcceptanceResolution
      summary: Read one visible post-acceptance resolution and typed owner states
      description: Returns closed step state and references only; evidence, Provider payloads and PII are never returned.
      parameters:
        - $ref: "#/components/parameters/PostAcceptanceResolutionId"
      responses:
        "200":
          description: Current resolution plan and explicit partial, unknown or terminal owner states
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAcceptanceResolutionResource"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/DependencyUnavailable" }
  /support/post-acceptance-resolutions/{resolutionId}/executions:
    post:
      tags: [Support]
      operationId: executePostAcceptanceResolution
      summary: Consume the exact approval and advance owner-specific resolution steps
      description: |
        Rechecks the assigned executor, persistent permissions, active Case/order
        relationship, exact request/revision/Order versions and current owner fact.
        Support claims one durable step before invoking its owner outside the Support
        transaction. Owner success is not presented until the following Support Audit
        and state commit succeeds. Provider timeout stays UNKNOWN/RECONCILING; a later
        recovery performs lookup and never assumes success or silently reissues work.
      parameters:
        - $ref: "#/components/parameters/PostAcceptanceResolutionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutePostAcceptanceResolutionRequest"
      responses:
        "200":
          description: Latest durable partial, unknown, manual-review or resolved state
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAcceptanceResolutionResource"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/DependencyUnavailable" }
  /support/post-acceptance-resolutions/{resolutionId}/reconciliations:
    post:
      tags: [Support]
      operationId: reconcilePostAcceptanceResolution
      summary: Explicitly reconcile a manual-review Payment refund by owner lookup
      description: |
        Initial S80 permits only PAYMENT_REFUND. The exact resolution/Order versions
        and idempotency payload are rechecked; POINT, COUPON, SETTLEMENT and notification
        reissue through this operator endpoint is rejected as unsafe.
      parameters:
        - $ref: "#/components/parameters/PostAcceptanceResolutionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconcilePostAcceptanceResolutionRequest"
      responses:
        "200":
          description: Latest durable reconciliation outcome
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAcceptanceResolutionResource"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/DependencyUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Unique within actor ID and API operation
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    OrderId:
      in: path
      name: orderId
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
    SupportCaseId:
      in: path
      name: caseId
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
    PostAcceptanceResolutionId:
      in: path
      name: resolutionId
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
    SupportOrderTimelineCaseId:
      in: query
      name: caseId
      required: true
      description: SupportCase that currently has the target Order linked as RELATED_ORDER.
      schema:
        $ref: "#/components/schemas/Identifier"
    SupportTimelineSources:
      in: query
      name: sources
      required: false
      description: Closed source filter. Omission selects every source valid for the endpoint.
      style: form
      explode: true
      schema:
        type: array
        uniqueItems: true
        maxItems: 9
        items:
          $ref: "#/components/schemas/SupportTimelineSource"
    SupportTimelineTypes:
      in: query
      name: types
      required: false
      description: Closed fact-type filter. Omission selects every type valid for the endpoint.
      style: form
      explode: true
      schema:
        type: array
        uniqueItems: true
        maxItems: 15
        items:
          $ref: "#/components/schemas/SupportTimelineType"
    SupportTimelineCursor:
      in: query
      name: cursor
      required: false
      description: |
        Opaque HMAC-signed cursor bound to the exact endpoint, Case/Order scope and
        canonical source/type filters. It expires after 15 minutes; malformed,
        expired or scope-mismatched values return 400 INVALID_REQUEST.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    SupportTimelineLimit:
      in: query
      name: limit
      required: false
      description: Maximum globally merged facts. Defaults to 20 and may not exceed 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Request syntax or field validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Role, ownership, or store membership check failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource is not visible in the authenticated actor scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: State, capacity, duplicate resource, or idempotency payload conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    DependencyUnavailable:
      description: A required dependency is unavailable; no fallback result was used
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  headers:
    NoStore:
      required: true
      description: Sensitive support representation must not be stored by shared or browser caches.
      schema:
        type: string
        const: no-store
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 128
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, details]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        correlationId:
          type: string
          minLength: 1
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        field:
          type: string
        reason:
          type: string
          minLength: 1
...
    SupportTimelineSource:
      type: string
      enum: [SUPPORT, ORDERING, PAYMENT, LOYALTY, PROMOTION, FULFILLMENT, SETTLEMENT, NOTIFICATION, OPERATIONS]
    SupportTimelineType:
      type: string
      enum:
        - CASE_STATE
        - CASE_ASSIGNMENT
        - CASE_INTERACTION
        - CASE_NOTE
        - SUBJECT_LINK
        - ORDER_STATE
        - PAYMENT_STATE
        - REFUND_STATE
        - POINT_RESERVATION
        - COUPON_RESERVATION
        - PICKUP_RESERVATION
        - SETTLEMENT_ITEM
        - SETTLEMENT_ADJUSTMENT
        - NOTIFICATION_DELIVERY
        - OPERATION_AUDIT
    SupportTimelineState:
      type: string
      enum:
        - OPEN
        - PENDING_CUSTOMER
        - PENDING_STORE
        - ESCALATED
        - RESOLVED
        - CLOSED
        - ASSIGNED
        - INBOUND
        - OUTBOUND
        - INTERNAL
        - RECORDED
        - LINKED
        - UNLINKED
        - PENDING_PAYMENT
        - PAID
        - ACCEPTED
        - PREPARING
        - READY
        - COMPLETED
        - REJECTED
        - EXPIRED
        - CANCELLED
        - APPROVING
        - APPROVED
        - FAILED
        - UNKNOWN
        - RECONCILING
        - MANUAL_REVIEW
        - REQUESTED
        - PROCESSING
        - SUCCEEDED
        - RESERVED
        - USED
        - RELEASED
        - CONFIRMED
        - ITEM_CREATED
        - ADJUSTMENT_RECORDED
        - PENDING
        - RETRY_SCHEDULED
    SupportTimelineItem:
      type: object
      additionalProperties: false
      required: [itemId, source, type, state, summary, amountKrw, occurredAt]
      properties:
        itemId:
          $ref: "#/components/schemas/Identifier"
        source:
          $ref: "#/components/schemas/SupportTimelineSource"
        type:
          $ref: "#/components/schemas/SupportTimelineType"
        state:
          $ref: "#/components/schemas/SupportTimelineState"
        summary:
          type: string
          minLength: 3
          maxLength: 100
          description: Server-generated closed type/state label; never owner free text or PII.
        amountKrw:
          type: [integer, "null"]
          format: int64
          minimum: 0
        occurredAt:
          type: string
          format: date-time
    SupportTimelinePage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/SupportTimelineItem"
        nextCursor:
          type: [string, "null"]
          minLength: 1
          maxLength: 2048
...
    PostAcceptanceResolutionOutcome:
      type: string
      enum: [FULL_REFUND, PARTIAL_REFUND, NO_MONETARY_RESOLUTION, MANUAL_SETTLEMENT_REVIEW]
    PostAcceptanceResolutionResponsibility:
      type: string
      enum: [CUSTOMER, STORE, PLATFORM, SHARED, UNDETERMINED]
    PostAcceptanceResolutionState:
      type: string
      enum: [PLANNED, EXECUTING, PARTIALLY_RESOLVED, RECONCILING, RESOLVED, MANUAL_REVIEW]
    PostAcceptanceResolutionStepType:
      type: string
      enum: [PAYMENT_REFUND, POINT_RESTORATION, COUPON_RESTORATION, SETTLEMENT_ADJUSTMENT, CUSTOMER_NOTIFICATION]
    PostAcceptanceResolutionStepState:
      type: string
      enum: [PENDING, PROCESSING, RETRY_SCHEDULED, SUCCEEDED, NOT_REQUIRED, UNKNOWN, RECONCILING, MANUAL_REVIEW, BLOCKED]
    CreatePostAcceptanceResolutionRequest:
      type: object
      additionalProperties: false
      required: [requestId, revisionNumber, expectedRequestVersion, expectedOrderVersion, outcome, responsibility, cashRefundKrw, restorePoints, restoreCoupon, evidenceDigest]
      properties:
        requestId:
          $ref: "#/components/schemas/Identifier"
        revisionNumber:
          type: integer
          minimum: 1
        expectedRequestVersion:
          type: integer
          format: int64
          minimum: 0
        expectedOrderVersion:
          type: integer
          format: int64
          minimum: 0
        outcome:
          $ref: "#/components/schemas/PostAcceptanceResolutionOutcome"
        responsibility:
          $ref: "#/components/schemas/PostAcceptanceResolutionResponsibility"
        cashRefundKrw:
          type: integer
          format: int64
          minimum: 0
        restorePoints:
          type: boolean
        restoreCoupon:
          type: boolean
        settlementAdjustmentKrw:
          type: [integer, "null"]
          format: int64
          maximum: -1
        evidenceDigest:
          type: string
          pattern: "^[0-9a-f]{64}$"
      description: |
        FULL_REFUND/PARTIAL_REFUND require positive cashRefundKrw. Non-refund
        outcomes require zero cash and no benefit restoration. STORE/SHARED require
        an exact negative settlementAdjustmentKrw except MANUAL_SETTLEMENT_REVIEW.
        CUSTOMER/PLATFORM/UNDETERMINED forbid an automatic Settlement adjustment.
    ExecutePostAcceptanceResolutionRequest:
      type: object
      additionalProperties: false
      required: [expectedResolutionVersion, expectedRequestVersion, expectedOrderVersion]
      properties:
        expectedResolutionVersion:
          type: integer
          format: int64
          minimum: 0
        expectedRequestVersion:
          type: integer
          format: int64
          minimum: 0
        expectedOrderVersion:
          type: integer
          format: int64
          minimum: 0
    ReconcilePostAcceptanceResolutionRequest:
      type: object
      additionalProperties: false
      required: [stepType, expectedResolutionVersion, expectedOrderVersion]
      properties:
        stepType:
          type: string
          const: PAYMENT_REFUND
        expectedResolutionVersion:
          type: integer
          format: int64
          minimum: 0
        expectedOrderVersion:
          type: integer
          format: int64
          minimum: 0
    PostAcceptanceResolutionStepResource:
      type: object
      additionalProperties: false
      required: [type, state, attemptCount, resultReference, failureCode, nextAttemptAt, ownerState]
      properties:
        type:
          $ref: "#/components/schemas/PostAcceptanceResolutionStepType"
        state:
          $ref: "#/components/schemas/PostAcceptanceResolutionStepState"
        attemptCount:
          type: integer
          minimum: 0
        resultReference:
          type: [string, "null"]
          minLength: 1
          maxLength: 240
        failureCode:
          type: [string, "null"]
          pattern: "^[A-Z0-9_]{1,80}$"
        nextAttemptAt:
          type: [string, "null"]
          format: date-time
        ownerState:
          type: [string, "null"]
          description: Current Notification delivery state for CUSTOMER_NOTIFICATION; null for financial owner steps.
    PostAcceptanceResolutionResource:
      type: object
      additionalProperties: false
      required: [resolutionId, supportCaseId, requestId, revisionNumber, orderId, triggerOrderState, triggerOrderVersion, outcome, responsibility, cashRefundKrw, restorePoints, restoreCoupon, settlementAdjustmentKrw, state, version, createdAt, updatedAt, steps]
      properties:
        resolutionId:
          $ref: "#/components/schemas/Identifier"
        supportCaseId:
          $ref: "#/components/schemas/Identifier"
        requestId:
          $ref: "#/components/schemas/Identifier"
        revisionNumber:
          type: integer
          minimum: 1
        orderId:
          $ref: "#/components/schemas/Identifier"
        triggerOrderState:
          type: string
          enum: [PREPARING, READY, COMPLETED]
        triggerOrderVersion:
          type: integer
          format: int64
          minimum: 0
        outcome:
          $ref: "#/components/schemas/PostAcceptanceResolutionOutcome"
        responsibility:
          $ref: "#/components/schemas/PostAcceptanceResolutionResponsibility"
        cashRefundKrw:
          type: integer
          format: int64
          minimum: 0
        restorePoints:
          type: boolean
        restoreCoupon:
          type: boolean
        settlementAdjustmentKrw:
          type: [integer, "null"]
          format: int64
          maximum: -1
        state:
          $ref: "#/components/schemas/PostAcceptanceResolutionState"
        version:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        steps:
          type: array
          minItems: 5
          maxItems: 5
          items:
            $ref: "#/components/schemas/PostAcceptanceResolutionStepResource"
...